Ensuring that the tests for rescale() test using integers with dtype int

+++ tests/test_transition_solver.py
@@ -99,7 +99,7 @@ def test_check_for_zero_units_bad():
 
 @patch.object(TransitionSolver, "__abstractmethods__", set())
 def test_rescale_rescaled_numpy():
-    A = np.ones((2, 2))
+    A = np.ones((2, 2)).astype(int)
     expected = np.array([[0.5, 0.5], [0.5, 0.5]])
     ts = TransitionSolver()
     np.testing.assert_array_equal(ts._rescale(A), expected)  # pylint: disable=protected-access
@@ -109,7 +109,7 @@ def test_rescale_rescaled_numpy():
 def test_rescale_rescaled_pandas():
     import pandas
 
-    a_df = pandas.DataFrame(np.ones((2, 2)), columns=["A", "B"])
+    a_df = pandas.DataFrame(np.ones((2, 2)), columns=["A", "B"]).astype(int)
     expected_df = pandas.DataFrame([[0.5, 0.5], [0.5, 0.5]], columns=["A", "B"])
     ts = TransitionSolver()
     np.testing.assert_array_equal(ts._rescale(a_df), expected_df)  # pylint: disable=protected-access
